Create homePage Component,ResourceList
